no more bigdecimal

+++ micro-streaming-analytics/consumer/src/main/java/com/rogerserra/model/Machine.java
@@ -15,5 +15,5 @@ public class Machine {
     private String name;
     private String result;
     //private LocalDateTime timestamp;
-    private BigDecimal energyConsumption;
+    private double energyConsumption;
 }
